adjust block/latest respose
From what I see on the devnet and the testnet, the new endpoint `block/latest` doesn't return an array, so we should fix the response structure here if we want to use it before the fallback. Otherwise, we will get two calls to the API, and fallback will be used anyway.